テスト駆動開発でGO言語を学びましょう/structs/structs_test.go
package main

import "testing"

/*
	func TestPerimeter(t *testing.T) {
		rectangle := Rectangle{10.0, 10.0}
		got := Perimeter(rectangle)
		want := 40.0

		if got != want {
			t.Errorf("got %.2f want %.2f", got, want)
		}
	}

	func TestArea(t *testing.T) {
		rectangle := Rectangle{12.0, 6.0}
		got := Area(rectangle)
		want := 72.0

		if got != want {
			t.Errorf("got %.2f want %.2f", got, want)
		}
	}

func TestArea(t *testing.T) {

		t.Run("rectangles", func(t *testing.T) {
			rectangle := Rectangle{12.0, 6.0}
			got := Area(rectangle)
			want := 72.0

			if got != want {
				t.Errorf("got %g want %g", got, want)
			}

		})
		t.Run("circles", func(t *testing.T) {
			circle := Circle{10.0}
			got := CircleCalculator(circle)
			//want := 0.0
			want := 314.1592653589793

			if got != want {
				t.Errorf("got %g want %g", got, want)
			}
		})
	}

func TestArea(t *testing.T) {

	t.Run("rectangles", func(t *testing.T) {
		rectangle := Rectangle{12, 6}
		got := rectangle.Area()
		want := 72.0

		if got != want {
			t.Errorf("got %g want %g", got, want)
		}
	})

	t.Run("circles", func(t *testing.T) {
		circle := Circle{10}
		got := circle.Area()
		want := 314.1592653589793

		if got != want {
			t.Errorf("got %g want %g", got, want)
		}
	})

}
*/
func TestArea(t *testing.T) {

	checkArea := func(t *testing.T, shape Shape, want float64) {
		t.Helper()
		got := shape.Area()
		if got != want {
			t.Errorf("got %g want %g", got, want)
		}
	}

	t.Run("rectangles", func(t *testing.T) {
		rectangle := Rectangle{12, 6}
		checkArea(t, rectangle, 72.0)
	})

	t.Run("circles", func(t *testing.T) {
		circle := Circle{10}
		checkArea(t, circle, 314.1592653589793)
	})

}
